pyproject.toml
[tool.poetry]
name = "shift-course"
version = "0.1.0"
description = ""
authors = ["Yana Shurinova <yana.shurinova009@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wemake-python-styleguide = "^0.18.0"
fastapi = "^0.104.0"
aiohttp = "^3.8.6"
pytest = "^7.4.2"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.1"
pytest-coverage = "^0.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
pydantic = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
